+++ gamesformykids/lib/utils/game/gameUtils.ts
@@ -1,37 +1,28 @@
 /**
  * אוסף פונקציות עזר גנריות למשחקים
  */
-import { FEEDBACK_MESSAGES, GAME_CONSTANTS, MEMORY_GAME_CONSTANTS, LETTER_HEBREW_PRONUNCIATIONS } from "../../constants";
-import { AUDIO_CONSTANTS } from "../../constants/core";
-import { speakHebrew, cancelSpeech, isSpeechEnabled } from '../speech/enhancedSpeechUtils';
-import { logError, logWarning } from '../errorUtils';
+import { GAME_CONSTANTS, LETTER_HEBREW_PRONUNCIATIONS } from "../../constants";
 import { useGameProgressStore } from '@/lib/stores';
+import { shuffleArray } from './cardUtils';
+import { speakPositiveFeedback, speakNegativeFeedback, speakStartMessage } from './feedbackUtils';
+
+// ── Re-exports from domain files (backward compat) ──────────────────────────
+export { playSuccessSound, playCustomSound, playAnimalSound, playMemorySuccessSound } from './audioUtils';
+export { shuffleArray, shuffle, createShuffledMemoryCards } from './cardUtils';
+export {
+  getRandomFeedbackMessage,
+  speakPositiveFeedback,
+  speakNegativeFeedback,
+  speakStartMessage,
+  speakItemName,
+} from './feedbackUtils';
 
 /**
  * פונקציית עזר להשהייה
- * @param ms זמן השהייה במילישניות
- * @returns Promise שמסתיים אחרי הזמן שהוגדר
  */
 export function delay(ms: number): Promise<void> {
   return new Promise(resolve => setTimeout(resolve, ms));
 }
-
-/**
- * פונקציית עזר - מערבבת מערך בצורה אקראית (Fisher-Yates)
- * @param array המערך לערבוב
- * @returns מערך חדש מעורבב
- */
-export function shuffleArray<T>(array: T[]): T[] {
-  const a = [...array];
-  for (let i = a.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [a[i], a[j]] = [a[j], a[i]];
-  }
-  return a;
-}
-
-/** קיצור לייבוא נוח בקבצי משחק */
-export const shuffle: <T>(arr: T[]) => T[] = shuffleArray;
 
 /**
  * מספר שלם אקראי בטווח [min, max] (כולל קצוות)
@@ -42,8 +33,6 @@ export function randInt(min: number, max: number): number {
 
 /**
  * מחזיר את ההגייה העברית של שם האות
- * @param letterName שם האות באנגלית
- * @returns הגייה עברית של האות
  */
 export function getHebrewPronunciation(letterName: string): string {
   return LETTER_HEBREW_PRONUNCIATIONS[letterName] || letterName;
@@ -51,315 +40,71 @@ export function getHebrewPronunciation(letterName: string): string {
 
 /**
  * בוחר פריט אקראי מתוך מערך
- * @param items מערך של אפשרויות
- * @returns פריט אקראי מהמערך
  */
 export function getRandomItem<T>(items: T[]): T {
   if (!items || items.length === 0) {
     throw new Error("Cannot get random item from empty array");
   }
-  
-  const randomIndex = Math.floor(Math.random() * items.length);
-  return items[randomIndex];
-}
-
-/**
- * בוחר הודעת משוב אקראית מרשימת הודעות
- * @param type סוג המשוב - הצלחה, טעות או התחלה
- * @returns טקסט אקראי מסוג המשוב שנבחר
- */
-export function getRandomFeedbackMessage(type: 'SUCCESS' | 'WRONG' | 'START'): string {
-  return getRandomItem(FEEDBACK_MESSAGES[type]);
-}
-
-/**
- * משמיע משוב קולי חיובי
- * @returns Promise שמסתיים עם סיום הדיבור
- */
-export async function speakPositiveFeedback(): Promise<void> {
-  cancelSpeech();
-  await delay(GAME_CONSTANTS.DELAYS.SUCCESS_SPEAK_DELAY);
-  try {
-    await speakHebrew(getRandomFeedbackMessage('SUCCESS'));
-  } catch (error) {
-    logError("שגיאה בהשמעת משוב חיובי:", error);
-  }
-}
-
-/**
- * משמיע משוב קולי שלילי
- * @returns Promise שמסתיים עם סיום הדיבור
- */
-export async function speakNegativeFeedback(): Promise<void> {
-  cancelSpeech();
-  await delay(GAME_CONSTANTS.DELAYS.WRONG_ANSWER_DELAY);
-  try {
-    await speakHebrew(getRandomFeedbackMessage('WRONG'));
-  } catch (error) {
-    logError("שגיאה בהשמעת משוב שלילי:", error);
-  }
-}
-
-/**
- * משמיע ברכת התחלה אקראית
- * @returns Promise שמסתיים עם סיום הדיבור
- */
-export async function speakStartMessage(): Promise<void> {
-  cancelSpeech();
-  await delay(GAME_CONSTANTS.DELAYS.START_GAME_DELAY);
-  try {
-    await speakHebrew(getRandomFeedbackMessage('START'));
-  } catch (error) {
-    logError("שגיאה בהשמעת ברכת התחלה:", error);
-  }
-}
-
-/**
- * משמיע את השם של פריט במשחק (אות, צבע, צורה וכו')
- * פונקציה גנרית שיכולה לשמש לכל סוגי המשחקים
- * @param itemName שם הפריט באנגלית
- * @param translator פונקציה שמתרגמת את השם לעברית
- * @returns Promise שמסתיים עם סיום הדיבור
- */
-export async function speakItemName(itemName: string, translator: (name: string) => string): Promise<void> {
-  if (!isSpeechEnabled()) {
-    return;
-  }
-
-  // מבטל כל דיבור קודם
-  cancelSpeech();
-
-  // השהייה קצרה לפני הכרזה
-  await delay(GAME_CONSTANTS.DELAYS.SPEAK_DELAY);
-
-  try {
-    const hebrewName = translator(itemName);
-    
-    const success = await speakHebrew(hebrewName);
-    if (!success) {
-      logWarning("Failed to speak item:", itemName);
-    }
-  } catch (error) {
-    logError("שגיאה בהשמעת הפריט:", error);
-  }
-}
-
-/**
- * פונקציה ליצירת צליל הצלחה - אקורד דו מז'ור
- * @param audioContext AudioContext להשמעת הצליל
- */
-export function playSuccessSound(audioContext: AudioContext | null): void {
-  if (!audioContext) return;
-  
-  // תדרים של צלילי אקורד דו מז'ור (C5, E5, G5)
-  const notes = AUDIO_CONSTANTS.SUCCESS_CHORD;
-  
-  notes.forEach((frequency, index) => {
-    // יצירת מחולל צליל וווליום
-    const oscillator = audioContext.createOscillator();
-    const gainNode = audioContext.createGain();
-    
-    // הגדרות הצליל
-    oscillator.type = "sine";
-    oscillator.frequency.value = frequency;
-    
-    // חיבור רכיבי השמע
-    oscillator.connect(gainNode);
-    gainNode.connect(audioContext.destination);
-
-    // תזמון הצלילים באקורד (בהפרשים קטנים)
-    const startTime = audioContext.currentTime + index * 0.1;
-    
-    // עקומת עוצמת הצליל
-    gainNode.gain.setValueAtTime(0, startTime);
-    gainNode.gain.linearRampToValueAtTime(0.1, startTime + 0.05);
-    gainNode.gain.exponentialRampToValueAtTime(0.01, startTime + 0.3);
-
-    // הפעלה והפסקה של הצליל
-    oscillator.start(startTime);
-    oscillator.stop(startTime + 0.3);
-  });
+  return items[Math.floor(Math.random() * items.length)];
 }
 
 /**
  * פונקציה גנרית ליצירת אפשרויות לבחירה במשחק עם תשובה נכונה אחת
- * @param correctItem האיבר הנכון שצריך להיות בין האפשרויות
- * @param allItems כל האיברים האפשריים
- * @param count מספר האפשרויות לבחירה
- * @param idField שם השדה שמשמש לזיהוי (ברירת מחדל 'name')
- * @returns מערך של אפשרויות מעורבבות, כולל התשובה הנכונה
  */
 export function generateOptions<T>(
-  correctItem: T, 
-  allItems: T[], 
+  correctItem: T,
+  allItems: T[],
   count: number = 4,
   idField: keyof T = 'name' as keyof T
 ): T[] {
-  // סינון האיברים השגויים (שאינם התשובה הנכונה)
   const incorrectItems = allItems.filter(item => item[idField] !== correctItem[idField]);
-  
-  // בחירת איברים שגויים אקראיים
   const selectedIncorrect = shuffleArray(incorrectItems).slice(0, count - 1);
-  
-  // ערבוב התשובות כולל התשובה הנכונה
   return shuffleArray([correctItem, ...selectedIncorrect]);
 }
 
 /**
- * פונקציה גנרית לטיפול במצב משחק
- * מעדכנת ניקוד ורמה דרך Zustand store, מפעילה חגיגה ומעבירה לשלב הבא
- * @param setShowCelebration callback לעדכון מצב חגיגה
- * @param onLevelComplete פונקציה שתופעל לאחר השלמת הרמה
+ * פונקציה גנרית לטיפול בתשובה נכונה
  */
 export async function handleCorrectGameAnswer(
   setShowCelebration: (v: boolean) => void,
   onLevelComplete: () => Promise<void>
 ): Promise<void> {
-  // עדכון ניקוד דרך store + הפעלת חגיגה
   useGameProgressStore.getState().incrementScore(GAME_CONSTANTS.SCORE_INCREMENT);
   setShowCelebration(true);
-  
-  // משוב קולי חיובי
+
   await speakPositiveFeedback();
-  
-  // השהייה אחרי המשוב
+
   await delay(GAME_CONSTANTS.DELAYS.NEXT_ITEM_DELAY);
   await delay(GAME_CONSTANTS.DELAYS.CELEBRATION_DURATION);
-  
-  // עדכון רמה דרך store + סגירת חגיגה
+
   useGameProgressStore.getState().incrementLevel(Infinity);
   setShowCelebration(false);
-  
-  // בחירת פריט חדש
+
   await onLevelComplete();
 }
 
 /**
  * פונקציה גנרית להתחלת משחק חדש
- * @param initialState מצב התחלתי של המשחק
- * @param setGameState פונקציה לעדכון מצב המשחק
- * @param onGameStart פונקציה שתופעל לאחר התחלת המשחק
  */
 export async function startNewGame<T>(
   initialState: T,
   setGameState: React.Dispatch<React.SetStateAction<T>>,
   onGameStart: () => Promise<void>
 ): Promise<void> {
-  // עדכון מצב המשחק למצב התחלתי
   setGameState(initialState);
-  
-  // השמעת ברכת התחלה
   await speakStartMessage();
-  
-  // השהייה לפני תחילת המשחק
   await delay(GAME_CONSTANTS.DELAYS.NEXT_ITEM_DELAY);
-  
-  // התחלת המשחק
   await onGameStart();
 }
 
 /**
  * פונקציה גנרית לטיפול בתשובה שגויה
- * משמיעה משוב שלילי ואז מאפשרת השמעת הפריט שוב
- * @param onSpeakItem פונקציה להשמעת שם הפריט הנוכחי
  */
 export async function handleWrongGameAnswer(
   onSpeakItem: () => Promise<void>
 ): Promise<void> {
-  // משמיע משוב שלילי
   await speakNegativeFeedback();
-  
-  // השהייה לפני חזרה על שם הפריט
   await delay(GAME_CONSTANTS.DELAYS.RETRY_DELAY);
-  
-  // חזרה על שם הפריט
   await onSpeakItem();
-}
-
-/**
- * פונקציה גנרית להשמעת צליל במשחק הזיכרון
- * @param audioContext אובייקט AudioContext להשמעת הצליל
- * @param frequencies מערך של תדרים להשמעה
- * @param type סוג הגל (sine, triangle וכו')
- * @param gainValue עוצמת הצליל
- * @param durationMs משך הצליל במילישניות
- */
-export function playCustomSound(
-  audioContext: AudioContext | null,
-  frequencies: number[],
-  type: OscillatorType = 'sine',
-  gainValue: number = 0.3,
-  durationMs: number = 150
-): void {
-  if (!audioContext) return;
-  
-  frequencies.forEach((freq, i) => {
-    const osc = audioContext.createOscillator();
-    const gain = audioContext.createGain();
-    
-    osc.connect(gain);
-    gain.connect(audioContext.destination);
-    
-    const t = audioContext.currentTime + i * 0.2;
-    osc.frequency.setValueAtTime(freq, t);
-    osc.type = type;
-    
-    gain.gain.setValueAtTime(0, t);
-    gain.gain.linearRampToValueAtTime(gainValue, t + 0.05);
-    gain.gain.exponentialRampToValueAtTime(0.01, t + (durationMs / 1000));
-    
-    osc.start(t);
-    osc.stop(t + (durationMs / 1000));
-  });
-}
-
-/**
- * פונקציה להשמעת צליל של חיה במשחק הזיכרון
- * @param audioContext אובייקט AudioContext להשמעת הצליל
- * @param emoji האימוג'י של החיה
- * @param frequencies אובייקט המיפוי של תדרי החיות
- */
-export function playAnimalSound(
-  audioContext: AudioContext | null,
-  emoji: string,
-  frequencies: Record<string, number[]>
-): void {
-  const animalFreqs = frequencies[emoji] || frequencies['default'];
-  playCustomSound(audioContext, animalFreqs);
-}
-
-/**
- * פונקציה גנרית להשמעת צליל הצלחה במשחק הזיכרון
- * @param audioContext אובייקט AudioContext להשמעת הצליל
- * @param frequencies מערך של תדרים להשמעה
- */
-export function playMemorySuccessSound(
-  audioContext: AudioContext | null,
-  frequencies: number[] = MEMORY_GAME_CONSTANTS.SUCCESS_SOUND_FREQUENCIES
-): void {
-  playCustomSound(audioContext, frequencies, 'triangle', 0.2, 80);
-}
-
-/**
- * פונקציה גנרית לערבוב וייצור קלפים למשחק זיכרון
- * @param items מערך פריטים ליצירת זוגות (אימוג'ים, תמונות וכו')
- * @returns מערך של קלפים מעורבבים עם זוגות
- */
-export function createShuffledMemoryCards<T>(items: T[]): { id: number, item: T, isFlipped: boolean, isMatched: boolean }[] {
-  // יוצרים זוגות ומערבבים
-  const shuffledCards = [...items, ...items]
-    .map((item) => ({
-      id: Math.random(), // מזהה ייחודי אקראי
-      item,
-      isFlipped: false,
-      isMatched: false,
-    }))
-    .sort(() => Math.random() - 0.5);
-  
-  // עדכון מזהים רציפים אחרי הערבוב
-  return shuffledCards.map((card, index) => ({
-    ...card,
-    id: index
-  }));
 }
