+++ gamesformykids/lib/utils/game/index.ts
@@ -1,4 +1,7 @@
 // Game Utilities
 export * from './gameUtils';
+export * from './audioUtils';
+export * from './cardUtils';
+export * from './feedbackUtils';
 export * from './gameMetadata';
 export * from './gameNavigation';
